Restructured R scripts.
